requirements.txt
jupyter
scikit-image~=0.18.1
numpy~=1.19.5
tensorflow-gpu>=2.4.1
matplotlib